CMSProBaseGateway/src/main/webapp/app/entities/CMSProBaseService/logos/logos-update.component.ts
import { Component, Inject } from 'vue-property-decorator';

import { mixins } from 'vue-class-component';
import JhiDataUtils from '@/shared/data/data-utils.service';

import { numeric, required, minLength, maxLength, minValue, maxValue, decimal } from 'vuelidate/lib/validators';

import AlertService from '@/shared/alert/alert.service';
import { ILogos, Logos } from '@/shared/model/CMSProBaseService/logos.model';
import LogosService from './logos.service';

const validations: any = {
  logos: {
    fileName: {
      required,
    },
    fileType: {},
    content: {
      required,
    },
    createdBy: {
      required,
    },
    creationDate: {
      required,
    },
  },
};

@Component({
  validations,
})
export default class LogosUpdate extends mixins(JhiDataUtils) {
  @Inject('alertService') private alertService: () => AlertService;
  @Inject('logosService') private logosService: () => LogosService;
  public logos: ILogos = new Logos();
  public isSaving = false;
  public currentLanguage = '';

  beforeRouteEnter(to, from, next) {
    next(vm => {
      if (to.params.logosId) {
        vm.retrieveLogos(to.params.logosId);
      }
    });
  }

  created(): void {
    this.currentLanguage = this.$store.getters.currentLanguage;
    this.$store.watch(
      () => this.$store.getters.currentLanguage,
      () => {
        this.currentLanguage = this.$store.getters.currentLanguage;
      }
    );
  }

  public save(): void {
    this.isSaving = true;
    if (this.logos.id) {
      this.logosService()
        .update(this.logos)
        .then(param => {
          this.isSaving = false;
          this.$router.go(-1);
          const message = this.$t('cmsProBaseGatewayApp.cmsProBaseServiceLogos.updated', { param: param.id });
          this.alertService().showAlert(message, 'info');
        });
    } else {
      this.logosService()
        .create(this.logos)
        .then(param => {
          this.isSaving = false;
          this.$router.go(-1);
          const message = this.$t('cmsProBaseGatewayApp.cmsProBaseServiceLogos.created', { param: param.id });
          this.alertService().showAlert(message, 'success');
        });
    }
  }

  public retrieveLogos(logosId): void {
    this.logosService()
      .find(logosId)
      .then(res => {
        this.logos = res;
      });
  }

  public previousState(): void {
    this.$router.go(-1);
  }

  public initRelationships(): void {}
}
