CMSProBaseGateway/src/test/javascript/spec/app/entities/CMSProBaseService/logos/logos-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import AlertService from '@/shared/alert/alert.service';
import * as config from '@/shared/config/config';
import LogosUpdateComponent from '@/entities/CMSProBaseService/logos/logos-update.vue';
import LogosClass from '@/entities/CMSProBaseService/logos/logos-update.component';
import LogosService from '@/entities/CMSProBaseService/logos/logos.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});

describe('Component Tests', () => {
  describe('Logos Management Update Component', () => {
    let wrapper: Wrapper<LogosClass>;
    let comp: LogosClass;
    let logosServiceStub: SinonStubbedInstance<LogosService>;

    beforeEach(() => {
      logosServiceStub = sinon.createStubInstance<LogosService>(LogosService);

      wrapper = shallowMount<LogosClass>(LogosUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          alertService: () => new AlertService(store),
          logosService: () => logosServiceStub,
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: '123' };
        comp.logos = entity;
        logosServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(logosServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.logos = entity;
        logosServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(logosServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });
  });
});
